Introduce PlaintextPassword
PlaintextPassword is a deserializable tuple struct with only
a String inside. There is no way to create this type other
than deserializing it from somewhere, and it can only be used
to create or verify a Password.

It does not implement Clone, so there's no way to share it
around more than once, and it must be consumed in the process
of creating or verifying a Password.